examples/neuro_react/pyproject.toml
[project]
name = "neuro-tags"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12.0"
dependencies = [
    "asyncord",
    "openai>=1.54.4",
    "python-dotenv>=1.0.1",
]

[tool.uv.sources]
asyncord = { path = "../../" }
